Learned about the SerializedField attribute
Learned that [ ] before a variable, class, or a method means we're adding an attribute to it. SerializedField lets us access the variable with this attribute in the inspector. The value in the inspector overrides the value for that variable in the code. The value in the inspector is saved to the disk.