Use Time.deltaTime
Time.deltaTime gives the time it took that last frame to run. Update runs every frame so if a computer was faster than another then the car would move faster in faster computer. Since we want to control the speed of the player across different kinds of computers we need to multiply Time.deltaTime to the steerAmount and moveAmount. This makes it framerate independent. If a computer runs at 100 frames per second, then Time.deltaTime would return 0.01 seconds. So 100 * 0.01 would return 1. A computer that runs at 10 frames would return 0.1 seconds and multiplying these together would also return 1. And since 1 times any number is that number, this makes it frame rate independent.