Use Late Update for Camera Follow
Learned that late update is used because it runs every frame but after everything in update has run. This make sure that the camera is updating to the player's current movements. If both were in the update function, it would be inconsistent whether the camera follow would run first or the player movement would. If the camera follow ran first then the camera would be using the player's previous position and not the current one and then the player would move. This would cause jitter in the camera movement.